+++ integrations/astro/modules/secrets.js
@@ -1,0 +1,4 @@
+export const getSecret = () =>
+	console.log(
+		"[CloudCannon] getSecret is not supported in an editable component. Please use an editing fallback instead.",
+	);