+++ nodes/editable-component.ts
@@ -19,6 +19,7 @@ export default class EditableComponent extends Editable {
 
 	private updatePromise: Promise<void> | undefined;
 	private needsReupdate = false;
+	private pendingPartialSubtree?: ChildNode | null;
 
 	getComponents() {
 		return getEditableComponentRenderers();
@@ -51,22 +52,41 @@ export default class EditableComponent extends Editable {
 		return true;
 	}
 
-	update(): Promise<void> {
+	update(partialSubtree?: ChildNode | null): Promise<void> {
 		if (this.updatePromise) {
 			this.needsReupdate = true;
+			this.pendingPartialSubtree = partialSubtree;
 			return this.updatePromise;
 		}
-		this.updatePromise = this._update().then(() => {
+		this.updatePromise = this._update(partialSubtree).then(() => {
 			this.updatePromise = undefined;
 			if (this.needsReupdate) {
 				this.needsReupdate = false;
-				return this.update();
+				const savedPartialSubtree = this.pendingPartialSubtree;
+				this.pendingPartialSubtree = undefined;
+				return this.update(savedPartialSubtree);
 			}
 		});
 		return this.updatePromise;
 	}
 
-	async _update(): Promise<void> {
+	santiseComponentOutput(el: HTMLElement): HTMLElement {
+		el.querySelectorAll("noscript").forEach((el) => el.remove());
+		return el;
+	}
+
+	async _update(partialSubtree?: ChildNode | null): Promise<void> {
+		if (partialSubtree) {
+			if (this.controlsElement) {
+				this.controlsElement.remove();
+			}
+			this.updateTree(this.element, partialSubtree);
+			if (this.controlsElement) {
+				this.element.appendChild(this.controlsElement);
+			}
+			return;
+		}
+
 		this.element.classList.remove("errored");
 
 		const key = this.element.dataset.component;
@@ -118,7 +138,7 @@ export default class EditableComponent extends Editable {
 
 		let rootEl: HTMLElement;
 		try {
-			rootEl = await component(value);
+			rootEl = this.santiseComponentOutput(await component(value));
 		} catch (err: unknown) {
 			this.element.classList.add("errored");
 			const error = document.createElement("editable-region-error-card");
@@ -151,99 +171,112 @@ export default class EditableComponent extends Editable {
 		targetNode?: ChildNode | null,
 		renderNode?: ChildNode | null,
 	): void {
-		let targetChild: ChildNode | null | undefined =
+		let nextTargetChild: ChildNode | null | undefined =
 			targetNode?.firstChild ?? undefined;
-		let renderChild: ChildNode | null | undefined =
+		let nextRenderChild: ChildNode | null | undefined =
 			renderNode?.firstChild ?? undefined;
-		while (renderChild || targetChild) {
-			const nextTargetChild: ChildNode | null | undefined =
-				targetChild?.nextSibling ?? undefined;
-			const nextRenderChild: ChildNode | null | undefined =
-				renderChild?.nextSibling ?? undefined;
 
-			if (
-				targetChild instanceof Element &&
-				renderChild &&
-				!(renderChild instanceof Element)
-			) {
+		while (nextRenderChild || nextTargetChild) {
+			const targetChild: ChildNode | null | undefined = nextTargetChild;
+			const renderChild: ChildNode | null | undefined = nextRenderChild;
+
+			nextTargetChild = targetChild?.nextSibling ?? undefined;
+			nextRenderChild = renderChild?.nextSibling ?? undefined;
+
+			// There is an existing node in the DOM but no rendered node in it's place
+			if (!renderChild && targetChild) {
+				targetNode?.removeChild(targetChild);
+				continue;
+			}
+
+			// There is a node to be rendered, but no existing node in the DOM in it's place
+			if (renderChild && !targetChild) {
+				targetNode?.appendChild(renderChild);
+				continue;
+			}
+
+			if (!targetChild || !renderChild) {
+				throw new Error("Illegal rendering state, both children should exist ");
+			}
+
+			// The existing child is an element and the rendered child is a node (i.e. some text)
+			if (targetChild instanceof Element && !(renderChild instanceof Element)) {
+				// Render the (text) node before the element and skip to the next rendered child
 				targetChild.before(renderChild);
-				renderChild = nextRenderChild;
+				nextTargetChild = targetChild;
 				continue;
 			}
 
-			if (
-				renderChild instanceof Element &&
-				targetChild &&
-				!(targetChild instanceof Element)
-			) {
+			// The existing child is a node (i.e. some text) and the rendered child is an element
+			if (renderChild instanceof Element && !(targetChild instanceof Element)) {
+				// Delete the (text) node and skip to the next existing child
 				targetChild.remove();
-				targetChild = nextTargetChild;
+				nextRenderChild = renderChild;
 				continue;
 			}
 
+			// Both existing and rendered children are nodes (i.e. some text)
 			if (
-				renderChild &&
-				targetChild &&
 				!(renderChild instanceof Element) &&
 				!(targetChild instanceof Element)
 			) {
+				// Render the offscreen node's content into the existing node.
 				targetChild.nodeValue = renderChild.nodeValue;
-			} else if (
-				targetChild instanceof HTMLElement &&
-				renderChild instanceof HTMLElement &&
-				isEditableElement(renderChild) &&
-				isEditableElement(targetChild)
-			) {
-				if (!areEqualEditables(renderChild, targetChild)) {
-					targetChild.replaceWith(renderChild);
-				} else if (isEditableText(renderChild) && isEditableText(targetChild)) {
-					if (
-						hasEditableText(targetChild) &&
-						!targetChild.editable.focused &&
-						!targetChild?.isEqualNode(renderChild) &&
-						hasEditable(renderChild)
-					) {
-						targetChild.replaceWith(renderChild);
-						for (let i = 0; i < this.listeners.length; i++) {
-							const listener = this.listeners[i];
-							if (listener.editable.element === targetChild) {
-								listener.editable.element = renderChild;
-								renderChild.editable.pushValue(
-									this.value,
-									this.specialProps,
-									listener,
-									{
-										...this.contexts,
-										__base_context: this.contextBase ?? {},
-									},
-								);
-							}
-						}
-					} else if (hasEditable(targetChild)) {
-						this.updateEditable(renderChild, targetChild);
-					}
-				} else if (hasEditable(targetChild)) {
-					this.updateEditable(renderChild, targetChild);
-				}
-			} else if (renderChild && targetChild) {
-				if (
-					renderChild.nodeName !== targetChild.nodeName ||
-					isEditableElement(renderChild) ||
-					isEditableElement(targetChild)
-				) {
-					targetChild.replaceWith(renderChild);
-				} else {
-					this.updateNode(targetChild, renderChild);
-					this.updateTree(targetChild, renderChild);
-				}
-			} else if (renderChild) {
-				targetNode?.appendChild(renderChild);
-			} else if (targetChild) {
-				targetNode?.removeChild(targetChild);
+				continue;
 			}
 
-			targetChild = nextTargetChild;
-			renderChild = nextRenderChild;
+			// Both children are elements but are different types
+			if (renderChild.nodeName !== targetChild.nodeName) {
+				targetChild.replaceWith(renderChild);
+				continue;
+			}
+
+			// Both existing and rendered children are the same kind of element, and neither is editable
+			if (!isEditableElement(renderChild) && !isEditableElement(targetChild)) {
+				// Update the existing element to match the rendered element and recurse their subtrees
+				this.updateNode(targetChild, renderChild);
+				this.updateTree(targetChild, renderChild);
+				continue;
+			}
+
+			if (
+				!(renderChild instanceof HTMLElement) ||
+				!(targetChild instanceof HTMLElement)
+			) {
+				throw new Error("Illegal state, both children should be elements");
+			}
+
+			// The existing and rendered child either aren't both editable or are different kinds of editables
+			if (!areEqualEditables(renderChild, targetChild)) {
+				targetChild.replaceWith(renderChild);
+				continue;
+			}
+
+			if (!isEditableElement(renderChild) || !isEditableElement(targetChild)) {
+				throw new Error("Illegal state, both children should be editable");
+			}
+
+			// Both elements are editable but the existing element hasn't hydrated
+			if (!hasEditable(targetChild)) {
+				targetChild.replaceWith(renderChild);
+				continue;
+			}
+
+			// The existing and rendered child are the same kind of non-text editable.
+			if (!isEditableText(targetChild)) {
+				this.updateEditable(renderChild, targetChild);
+				continue;
+			}
+
+			// The existing element is currently focused or is already the same as the rendered element
+			if (
+				(hasEditableText(targetChild) && targetChild.editable.focused) ||
+				targetChild?.isEqualNode(renderChild)
+			) {
+				this.updateEditable(renderChild, targetChild);
+			} else {
+				targetChild.replaceWith(renderChild);
+			}
 		}
 	}
 
@@ -303,6 +336,7 @@ export default class EditableComponent extends Editable {
 						...this.contexts,
 						__base_context: this.contextBase ?? {},
 					},
+					renderChild,
 				);
 			}
 		}
@@ -327,23 +361,6 @@ export default class EditableComponent extends Editable {
 				},
 			}),
 		);
-	}
-
-	setupListeners(): void {
-		super.setupListeners();
-		const key = this.element.dataset.component;
-		if (!key) {
-			return;
-		}
-
-		const component = this.getComponents()?.[key];
-		if (!component) {
-			document.addEventListener(
-				`editable-regions:registered-${key}`,
-				() => this.update(),
-				{ once: true },
-			);
-		}
 	}
 
 	mount(): void {